Move DebugReporter transformer to own file
