+++ src/AstTransforms.h
@@ -19,7 +19,7 @@
 #include "AstArgument.h"
 #include "AstTransformer.h"
 #include "AstTranslationUnit.h"
-#include "DebugReport.h"
+#include "DebugReporter.h"
 #include "Util.h"
 #include <functional>
 #include <memory>
